% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div_fcm.R
\name{beta_div_fcm}
\alias{beta_div_fcm}
\title{Beta diversity function for FCM data}
\usage{
beta_div_fcm(x, d = 3, n = 1, dist = "bray", k = 2, iter = 100,
  ord.type = c("NMDS", "PCoA"))
}
\arguments{
\item{x}{flowbasis object generated by flowBasis()}

\item{d}{Rounding factor for density values. Defaults to 4.}

\item{n}{Number of samples to pool? Defaults to 1. For n>1 samples will be pooled and averaged. 
Make sure samples are ordered per factor level.}

\item{dist}{Distance metric to use in the vegdist() function. Defaults to bray.}

\item{k}{Number of dimensions to project your samples into. Defaults to 2.}

\item{iter}{Number of iterations for NMDS analysis. Defaults to 100.}

\item{ord.type}{Choose between NMDS or PCoA analysis.}
}
\description{
This function performs beta diversity analysis on FCM data. 
The output can be fed into the plot.beta.fcm() function for visualization.
}
\keyword{beta_div_fcm(x)}
\keyword{betadiversity,}
\keyword{fcm}

