% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_Richness.R
\name{cum_Richness}
\alias{cum_Richness}
\title{Cumulative Richness function for FCM data}
\usage{
cum_Richness(x, d = 3)
}
\arguments{
\item{x}{flowbasis object generated by flowBasis()}

\item{d}{Rounding factor for density values. Defaults to 4.}
}
\description{
This function calculates Hill diversity metrics from FCM data
}
\examples{
cum_Richness(x)
}
\keyword{alpha}
\keyword{cumulative}
\keyword{fcm,}
\keyword{richness,}

