% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity.R
\name{Diversity}
\alias{Diversity}
\title{Diversity function for FCM data}
\usage{
Diversity(x, d = 4, plot = FALSE, R = 999)
}
\arguments{
\item{x}{flowbasis object generated by flowBasis()}

\item{d}{Rounding factor for density values. Defaults to 4.}

\item{plot}{Make plot of diversity values? Defaults to FALSE.}

\item{R}{Number of bootstraps to conduct. Defaults to 999}
}
\description{
This function calculates Hill diversity metrics from FCM data
}
\keyword{#}
\keyword{##}
\keyword{#'}
\keyword{'FL3-H','SSC-H','FSC-H')}
\keyword{'Total}
\keyword{(cells}
\keyword{(imported}
\keyword{<-}
\keyword{=}
\keyword{@examples}
\keyword{Asinh}
\keyword{Calculate}
\keyword{Cells')}
\keyword{Create}
\keyword{Define}
\keyword{Diversity(CoolingTower,}
\keyword{Diversity(fbasis,}
\keyword{Full}
\keyword{Gating}
\keyword{Green}
\keyword{I).}
\keyword{Isolate}
\keyword{Load}
\keyword{Normalize}
\keyword{PolygonGate}
\keyword{Short}
\keyword{Subset(flowData_transformed,}
\keyword{Sybr}
\keyword{[0,1]}
\keyword{`FL1-H`,}
\keyword{`FL3-H`=asinh(`FL3-H`),}
\keyword{`FL3-H`=mytrans(`FL3-H`),}
\keyword{`FSC-H`=asinh(`FSC-H`))}
\keyword{`FSC-H`=mytrans(`FSC-H`))}
\keyword{`SSC-H`=asinh(`SSC-H`),}
\keyword{`SSC-H`=mytrans(`SSC-H`),}
\keyword{a}
\keyword{across}
\keyword{alpha}
\keyword{and}
\keyword{apply(x,2,max),use.exprs=TRUE)}
\keyword{axis}
\keyword{axis.default,}
\keyword{based}
\keyword{bw=0.01,normalize=function(x)}
\keyword{c('FL1-H','FL3-H')}
\keyword{c(x,x,x,x,y,y,y,y)}
\keyword{cellular}
\keyword{cex=2),}
\keyword{check}
\keyword{colnames(sqrcut1)}
\keyword{coordinates}
\keyword{data}
\keyword{data(CoolingTower)}
\keyword{data(flowData)}
\keyword{data=flowData_transformed[1],}
\keyword{dataset}
\keyword{denoising}
\keyword{diversity}
\keyword{diversity,}
\keyword{example}
\keyword{fbasis}
\keyword{fcm,}
\keyword{filter=polyGate1,}
\keyword{filterId}
\keyword{fingerprint}
\keyword{flowBasis(flowData_transformed,}
\keyword{flowCore)}
\keyword{flowData_transformed}
\keyword{flowData_transformed[,param]}
\keyword{font=2,}
\keyword{for}
\keyword{format:}
\keyword{fsApply(x=flowData_transformed,FUN=function(x)}
\keyword{function(x)}
\keyword{gate}
\keyword{in}
\keyword{information}
\keyword{interest}
\keyword{interval}
\keyword{matrix(c(8.75,8.75,14,14,3,7.5,14,3),ncol=2,}
\keyword{max}
\keyword{max(summary[,1])}
\keyword{max.}
\keyword{mytrans}
\keyword{nbin=125,}
\keyword{nbin=128,}
\keyword{nrow=4)}
\keyword{object}
\keyword{of}
\keyword{on}
\keyword{only}
\keyword{par.strip.text=list(col='white',}
\keyword{param,}
\keyword{param=c('FL1-H',}
\keyword{parameter}
\keyword{parameters}
\keyword{plot=TRUE)}
\keyword{polyGate1}
\keyword{polyGate1)}
\keyword{polygonGate(.gate=sqrcut1,}
\keyword{precomputed}
\keyword{processing}
\keyword{quality}
\keyword{raw}
\keyword{scales=list(y=list(limits=c(0,14)),}
\keyword{select}
\keyword{smooth=FALSE)}
\keyword{sqrcut1}
\keyword{stained}
\keyword{summary}
\keyword{the}
\keyword{to}
\keyword{transform}
\keyword{transform(flowData,`FL1-H`=asinh(`FL1-H`),}
\keyword{transform(flowData_transformed,`FL1-H`=mytrans(`FL1-H`),}
\keyword{using}
\keyword{value}
\keyword{values}
\keyword{were}
\keyword{with}
\keyword{x)}
\keyword{x/max}
\keyword{x=list(limits=c(6,16))),}
\keyword{xyplot(`FL3-H`}
\keyword{~}

