% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_16S.R
\name{Diversity_16S}
\alias{Diversity_16S}
\title{Diversity function for 16S amplicon data}
\usage{
Diversity_16S(x, R = 999)
}
\arguments{
\item{x}{phyloseq object generated by the phyloseq package}

\item{R}{Number of bootstraps to conduct. Defaults to 999}
}
\description{
This function calculates Hill diversity metrics from 16S amplicon data (in phyloseq format).
D0 (richness) is calculated with three methods: 1) Observed richness, 2) Chao1 estimator
3) breakaway method (Willis and Bunge 2016). D1 (exponential of Shannon entropy)
and D2 (inverse Simpson index) are respectively Hill order 1 and 2.
Errors for D1 and D2 are calculated by bootstrapping.
}
\examples{
Diversity_16S(x)
}
\keyword{alpha}
\keyword{diversity,}
\keyword{fcm,}

