% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_gmm.R
\name{Diversity_gmm}
\alias{Diversity_gmm}
\title{Unsupervised calculation of phenotypic diversity on gaussian mixture-based phenotypic fingerprints}
\usage{
Diversity_gmm(gmm, R = 100, verbose = FALSE)
}
\arguments{
\item{gmm}{list containing 1) contigency table and 2) the gmm model created by 
the PhenoGMM function}

\item{R}{Number of bootstraps to conduct. Defaults to 100}

\item{verbose}{Progress of function is reported. Defaults to FALSE}
}
\description{
This function calculates Hill diversity metrics from the mixtures calculated by PhenoGMM.
D0 is calculated as Observed richness. D1 (exponential of Shannon entropy)
and D2 (inverse Simpson index) are respectively Hill order 1 and 2.
Errors for D0, D1 and D2 are calculated by bootstrapping.
}
\examples{
data(flowData_transformed)
# Make model on training data
testGMM <- PhenoGMM(flowData_transformed[1:2], downsample = 1e3, nG = 128, param = c("FL1-H", "FL3-H"))
# Apply model to unseen/new data
testPred <- PhenoMaskGMM(flowData_transformed, gmm = testGMM)
# Calculate diversity for both contigency tables
Diversity_gmm(testGMM)
Diversity_gmm(testPred)
}
\keyword{PhenoGMM}
\keyword{alpha,}
\keyword{diversity,}
\keyword{fcm,}
\keyword{gmm,}
