% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evenness.R
\name{Evenness}
\alias{Evenness}
\title{Evenness function for FCM data}
\usage{
Evenness(x, d = 3, n = 1, plot = FALSE)
}
\arguments{
\item{x}{flowbasis object generated by flowBasis()}

\item{d}{Rounding factor for density values. Defaults to 4.}

\item{n}{Number of replicates. Defaults to 1.}

\item{plot}{Make plot of diversity values? Defaults to FALSE.}
}
\description{
This function calculates Pareto evenness from FCM data
}
\examples{
Evenness()
}
\keyword{alpha}
\keyword{evenness,}
\keyword{fcm,}

