% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_csv.R
\name{export_csv}
\alias{export_csv}
\title{Exports the flow cytometric fingerprint or raw data to CSV.}
\usage{
export_csv(x, location = "./")
}
\arguments{
\item{x}{flowCore::flowFrame or flowFDA::flowBasis object}

\item{location}{absolute or relative path to where the resulting csv files 
are written to. If it does not exist, it will be created (provided the user
has write persmissions in that directory). Defaults to working directory.}
}
\description{
This function exports the fingerprint to a single CSV file or the fcs file 
associated with a flowFrame object to a CSV file .
}
\examples{

data(CoolingTower)
data(flowData)
### show class info to see if flowSet (collection of flowFrames) or flowBasis
class(flowData)
###  Export csv files of the raw data
fsApply(flowData,export_csv,location="./raw_csv/")
###  Export csv files of the fingerprint
export_csv(CoolingTower,location="./fingerprint/")
}
\keyword{fcm}
\keyword{learning,}
\keyword{machine}
