% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCS_clean.R
\name{FCS_clean}
\alias{FCS_clean}
\title{FCS_clean function for FCM data}
\usage{
FCS_clean(x, cleanparam = c(9, 11))
}
\arguments{
\item{x}{flowSet containing the samples to analyse. Cannot have been subsetted to 
a lower number of parameters.}

\item{cleanparam}{Indices of parameters to be used for removing errant observations. 
Defaults to 9 and 11 for FL1-H and FL3-H on the BD C6 Accuri FCM.}
}
\description{
This function denoises a flowSet object using the flowClean package. Observations that
do not need meet the criteria will be removed. Only works for samples with more than
30,000 observations.
}
\examples{
## Full data processing example
data(flowData)
cleanparam <- c(9,11)
flowData_cleaned <- flowCore::fsApply(x = flowData[1:5], FUN = function(x) FCS_clean(x, cleanparam))
}
\keyword{alpha}
\keyword{denoise,}
\keyword{fcm,}
