% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowData.R
\docType{data}
\name{flowData}
\alias{flowData}
\title{Flow cytometry test data of a cooling water microbial community}
\format{
A flowSet with 41 experiments
}
\source{
\url{http://dx.doi.org/10.5061/dryad.m1c04}
}
\usage{
flowData
}
\description{
Dataset containing the multiparametric flow cytometry results of 41 samples.
This is a subset of the total available dataset described in
doi: 10.1111/2041-210X.12607.
Available parameters: FSC-A SSC-A FL1-A FL2-A FL3-A FL4-A FSC-H SSC-H 
FL1-H FL2-H FL3-H FL4-H Width Time
}
\keyword{datasets}
