% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowData_ax.R
\docType{data}
\name{flowData_ax}
\alias{flowData_ax}
\title{Flow cytometry test data of 2 axenic bacterial strains}
\format{
A flowSet with 2 experiments
}
\source{
\url{http://flowrepository.org/id/FR-FCM-ZZSH}
}
\usage{
flowData_ax
}
\description{
Dataset containing the multiparametric flow cytometry results of 2 axenic bacterial strains
This is a subset of the total available dataset described in
doi:10.1371/journal.pone.0169754
Available parameters: FSC-A SSC-A FL1-A FL2-A FL3-A FL4-A FSC-H SSC-H 
FL1-H FL2-H FL3-H FL4-H Width Time
}
\keyword{datasets}
