% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoGMM.R
\name{PhenoGMM}
\alias{PhenoGMM}
\title{Train GMM-fitted model to FCS data.}
\usage{
PhenoGMM(fcs_x, param, downsample = 0, nG = 128)
}
\arguments{
\item{fcs_x}{flowSet object with input data on which the model should be built}

\item{param}{parameters to be used in the mixture modeling.}

\item{downsample}{Indicate to which sample size individual samples should be downsampled. 
By default no downsampling is performed}

\item{nG}{Number of mixtures to use. Defaults to 128.}
}
\description{
Train GMM-fitted model to FCS data.
}
\examples{
data(flowData_transformed)
testGMM <- PhenoGMM(flowData_transformed, downsample = 1e3, nG = 128, param = c("FL1-H", "FL3-H"))
testPred <- PhenoMaskGMM(flowData_transformed, gmm = testGMM)
}
\keyword{fingerprint}
