% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoMaskGMM.R
\name{PhenoMaskGMM}
\alias{PhenoMaskGMM}
\title{Apply PhenoGMM made model to new data.}
\usage{
PhenoMaskGMM(fcs_x, gmm)
}
\arguments{
\item{fcs_x}{flowSet object on which to apply the GMM mask}

\item{gmm}{model mask used to assign cluster allocations}
}
\description{
Apply PhenoGMM made model to new data.
}
\examples{
data(flowData_transformed)
testGMM <- PhenoGMM(flowData_transformed, downsample = 1e3, nG = 128, param = c("FL1-H", "FL3-H"))
testPred <- PhenoMaskGMM(flowData_transformed, gmm = testGMM)
}
\keyword{fingerprint}
