% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhTrans.R
\name{PhTrans}
\alias{PhTrans}
\title{Small function for normalizing channels based on maximum value of chosen
channel}
\usage{
PhTrans(channel, max_v)
}
\arguments{
\item{channel}{FCM channel}

\item{max_v}{maximum value across chosen FCM channel}
}
\description{
Small function for normalizing channels based on maximum value of chosen
channel
}
\keyword{fcm}
\keyword{max_v)}
\keyword{myTrans(`FL1-H`,}
\keyword{transformation,}
