% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq_test.R
\docType{data}
\name{physeq_test}
\alias{physeq_test}
\title{Microbial community composition based on the V3-V4 16S amplicon sequencing data 
from doi: 10.1111/2041-210X.12607. Data is stored as a Phyloseq object.}
\format{
phyloseq object
\describe{
  \item{"@otu_table"}{OTU table}
  \item{"@tax_table"}{Consensus taxonomy for each OTU}
}
}
\source{
\url{doi: 10.1111/2041-210X.12607}
}
\usage{
physeq_test
}
\description{
Microbial community composition based on the V3-V4 16S amplicon sequencing data 
from doi: 10.1111/2041-210X.12607. Data is stored as a Phyloseq object.
}
\keyword{datasets}
