% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.beta.fcm.R
\name{plot.beta.fcm}
\alias{plot.beta.fcm}
\title{Plot function for beta diversity analysis of FCM data}
\usage{
\method{plot}{beta.fcm}(x, color = NA, shape = NA, labels = c("Factor 1",
  "Factor 2"), legend.pres = NULL)
}
\arguments{
\item{x}{flowbasis object generated by flowBasis()}

\item{d}{Rounding factor for density values. Defaults to 4.}

\item{plot}{Make plot of diversity values? Defaults to FALSE.}

\item{R}{Number of bootstraps to conduct. Defaults to 999}
}
\description{
This function visualizes the beta diversity analysis by beta.div.fcm()
}
\examples{
plot.beta.fcm()
}
\keyword{alpha}
\keyword{diversity,}
\keyword{fcm,}

