% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_fcm.R
\name{plot_beta_fcm}
\alias{plot_beta_fcm}
\title{Plot function for beta diversity analysis of FCM data}
\usage{
plot_beta_fcm(x, color = NA, shape = NA, labels = c("Factor 1",
  "Factor 2"), legend.pres = NULL)
}
\arguments{
\item{x}{flowbasis object generated by flowBasis()}

\item{color}{A vector of factors indicating the groups to be colored in the plot}

\item{shape}{A vector of factors indicating the shape of the groups in the plot}

\item{labels}{A vector of length 2, giving the labels for color and shape to be 
used in the legend. Has to be in the order c("color","shape").}

\item{legend.pres}{Surpresses legend when there is no definition of labels.}
}
\description{
This function visualizes the beta diversity analysis generated by beta_div_fcm()
}
\examples{
plot_beta_fcm(x)
}
\keyword{beta}
\keyword{diversity,}
\keyword{fcm,}

