% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_FCS.R
\name{RandomF_FCS}
\alias{RandomF_FCS}
\title{Random Forest classifier for supervised demarcation of groups using flow cytometry data.}
\usage{
RandomF_FCS(x, sample_info, target_label, downsample = 0,
  classification_type = "sample", param = c("FL1-H", "FL3-H", "FSC-H",
  "SSC-H"), p_train = 0.75, seed = 777)
}
\arguments{
\item{x}{flowSet object where the necessary metadata for classification is included in the phenoData ac}

\item{sample_info}{Sample information necessary for the classification, has to contain a column named "name"
which matches the samplenames of the FCS files stored in the flowSet.}

\item{target_label}{Label that you should be predicted based on the flow cytometry data.}

\item{downsample}{Indicate to which sample size should be downsampled. 
By default samples are downsampled to the sample size of the sample with the lowest number of cells.}

\item{classification_type}{Whether classification is desired on "sample" level or on "single-cell" level. 
Defaults to sample level.}

\item{param}{Parameters to base classification on.}

\item{p_train}{Percentage of the data set that should be used for training the model.}

\item{seed}{Set random seed to be used during the analysis. Put at 777 by default.}
}
\description{
Random Forest classifier for supervised demarcation of groups using flow cytometry data.
}
\examples{
# Load raw data (imported using flowCore)
data(flowData)

# Format necessary metadata
metadata <- data.frame(names = flowCore::sampleNames(flowData), 
do.call(rbind, lapply(strsplit(flowCore::sampleNames(flowData),"_"), rbind)))
colnames(metadata) <- c("name", "Cycle_nr", "Location", "day", 
"timepoint", "Staining", "Reactor_phase", "replicate")

# Run Random Forest classifier to predict the Reactor phase based on the
# single-cell FCM data
RandomF_FCS(flowData, sample_info = metadata, target_label = "Reactor_phase",
downsample = 500)
}
\keyword{fcm}
\keyword{resampling,}
