% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/So.R
\name{So}
\alias{So}
\title{Structural organisation (So) for FCM data based on Koch et al. 2014}
\usage{
So(x, d = 3, n = 1, plot = FALSE)
}
\arguments{
\item{x}{flowbasis object generated by flowBasis()}

\item{d}{Rounding factor for density values. Defaults to 4.}

\item{n}{Number of replicates. Defaults to 1.}

\item{plot}{Make plot of diversity values? Defaults to FALSE.}
}
\description{
This function calculates Structural organisation from FCM data according to Koch et al. 2014.
}
\keyword{So(x)}
\keyword{So,}
\keyword{fcm,}
\keyword{organization}
\keyword{structural}

