% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_discretization.R
\name{time_discretization}
\alias{time_discretization}
\title{Time discretization function for FCM data}
\usage{
time_discretization(x, analysis.length, create = FALSE, start = 0,
  time.interval)
}
\arguments{
\item{x}{flowSet of FCM data}

\item{analysis.length}{Dataframe with $time defining the total analysis time for each sample}

\item{create}{Do you want to make a new folder for the new flowframes? Defaults to FALSE.}

\item{start}{Vector of length n(x) that indicates for each sample at what time point it should start discretisizing.
(e.g., first 10 minutes are irrelevant, start = 10*60)}

\item{time.interval}{Bin size of each new FCS file. For example, time.interval = 10 will make new FCS files of 10 second intervals.}
}
\description{
This function subsets the input frames based on analysis time. Developped for analysis of on-line time series data.
}
\examples{
# To be added in the near future
}
\keyword{FCM}
\keyword{analysis}
\keyword{fcm,}
\keyword{online,}
\keyword{series}
\keyword{time}

