% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip.R
\name{trip}
\alias{trip}
\title{Small (inefficient) function for taking mean/sd}
\usage{
trip(x, n = 3)
}
\arguments{
\item{x}{Input matrix}

\item{n}{Number of replicates}
}
\description{
Small (inefficient) function for taking mean/sd
}
\examples{
trip(x)
}
\keyword{fcm}
\keyword{resampling,}

