% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_col.R
\name{trip_col}
\alias{trip_col}
\title{Small (inefficient) function for extracting sample names associated with replicates}
\usage{
trip_col(x, n = 3)
}
\arguments{
\item{x}{Vector of Sample names of entire flowSet}

\item{n}{Number of replicates}
}
\description{
Small (inefficient) function for extracting sample names associated with replicates
}
\keyword{fcm}
\keyword{resampling,}
\keyword{trip()}

