% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\title{Calculating coefficient of variance (CV) of fingerprints from FCM data.}
\usage{
CV(x, d = 3, n = 1, plot = FALSE)
}
\arguments{
\item{x}{flowbasis object generated by flowBasis()according to Koch et al. 2014.}

\item{d}{Rounding factor for density values. Defaults to 4.}

\item{n}{Number of replicates. Defaults to 1.}

\item{plot}{Make plot of diversity values? Defaults to FALSE.}
}
\description{
This function calculates the coefficient of variance from the fingerprint generated by flowBasis().
}
\examples{
CV()

}
\keyword{CV,}
\keyword{coefficient}
\keyword{fcm,}
\keyword{of}
\keyword{variance}

