% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity.R
\name{Diversity}
\alias{Diversity}
\title{Diversity function for FCM data}
\usage{
Diversity(x, d = 4, plot = FALSE, R = 999)
}
\arguments{
\item{x}{flowbasis object generated by flowBasis()}

\item{d}{Rounding factor for density values. Defaults to 4.}

\item{plot}{Make plot of diversity values? Defaults to FALSE.}

\item{R}{Number of bootstraps to conduct. Defaults to 999}
}
\description{
This function calculates Hill diversity metrics from FCM data
}
\examples{
Diversity(x)
}
\keyword{alpha}
\keyword{diversity,}
\keyword{fcm,}

