% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCS_clean.R
\name{FCS_clean}
\alias{FCS_clean}
\title{FCS_clean function for FCM data}
\usage{
FCS_clean(
  x,
  param = c("FL1-H", "FL3-H", "FSC-H", "SSC-H"),
  timesplit = 0.1,
  TimeChannel = "Time"
)
}
\arguments{
\item{x}{flowSet containing the samples to analyse. 
Cannot have been subsetted to  a lower number of parameters.}

\item{param}{Parameters to base removing of errant observartions on.}

\item{timesplit}{Fraction of timestep used in flowAI for denoising. 
Please consult the `flowAI::flow_auto_qc` function for more information.}

\item{TimeChannel}{Name of time channel in the FCS files. This can differ 
between flow cytometers. Defaults to "Time". You can check this by: 
colnames(flowSet).}
}
\description{
This function denoises a flowSet object using the flowAI package. 
Observations that #' do not need meet the criteria will be removed. 
Only works for samples with more than 1000 observations.
}
\examples{
## Full data processing example
data(flowData)
flowData_cleaned <- FCS_clean(flowData[1:3])
}
\keyword{alpha}
\keyword{denoise,}
\keyword{fcm,}
