% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCS_resample.R
\name{FCS_resample}
\alias{FCS_resample}
\title{Resampling function for flowSet objects}
\usage{
FCS_resample(x, sample = 0, replace = FALSE)
}
\arguments{
\item{x}{flowSet object.}

\item{sample}{Desired sample size. Defaults to minimum sample size.}

\item{replace}{Do you want to resample with or without replacement? Defaults to FALSE, which is without replacement.}
}
\description{
This function resamples each sample to an equal number of cells
}
\keyword{fcm}
\keyword{resampling,}

