% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log.R
\name{Log}
\alias{Log}
\title{Small function for printing progress of parallel processes}
\usage{
Log(text, ...)
}
\arguments{
\item{text}{Input text}

\item{...}{additional parameters passed on to sprintf}
}
\description{
Small function for printing progress of parallel processes
}
\keyword{Log(x)}
\keyword{fcm}
\keyword{resampling,}
