% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoGMM.R
\name{PhenoGMM}
\alias{PhenoGMM}
\title{Apply existing GMM-fitted model to new FCS data. 
These data must contain the same parameters as those on which the model was constructed.}
\usage{
PhenoGMM(fcs_x, param, downsample = 0, nG = 128)
}
\arguments{
\item{fcs_x}{flowSet object with input data on which the model should be built}

\item{downsample}{Indicate to which sample size individual samples should be downsampled. 
By default no downsampling is performed}

\item{nG}{Number of mixtures to use. Defaults to 128.}
}
\description{
Apply existing GMM-fitted model to new FCS data. 
These data must contain the same parameters as those on which the model was constructed.
}
\examples{
testGMM <- PhenoGMM(flowData_transformed, downsample = 1e3, nG = 128, param = c("FL1-H", "FL3-H"))
testPred <- PhenoMaskGMM(flowData_transformed, gmm = testGMM)
}
\keyword{fingerprint}
