% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoMaskGMM.R
\name{PhenoMaskGMM}
\alias{PhenoMaskGMM}
\title{Create sample fingerprints from FCM data}
\usage{
PhenoMaskGMM(fcs_x, gmm)
}
\arguments{
\item{fcs_x}{flowSet object on which to apply the GMM mask}

\item{gmm}{model mask used to assign cluster allocations}
}
\description{
Create sample fingerprints from FCM data
}
\examples{
system.time(testx <- KytoGrid(flowData_transformed, param = c("FL1-H", "FL3-H")))
system.time(test <- flowFDA::flowBasis(flowData_transformed, param=c("FL1-H", "FL3-H"),
                                       nbin = 128, bw = 0.01, normalize = function(x) x))
#D2.boot <- function(x, i) 1/sum((x[i]/sum(x[i]))^2)
#D2_testx <- apply(testx, 1, FUN = D2.boot)
#D2_test <- apply(test@basis, 1, FUN = D2.boot)
#plot(D2_testx, D2_test)
}
\keyword{fingerprint}
