% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoMaskGMM.R
\name{PhenoMaskGMM}
\alias{PhenoMaskGMM}
\title{Apply PhenoGMM made model to new data.}
\usage{
PhenoMaskGMM(fcs_x, gmm, fcs_scale = FALSE)
}
\arguments{
\item{fcs_x}{flowSet object on which to apply the GMM mask}

\item{gmm}{model mask used to assign cluster allocations}

\item{fcs_scale}{Should data be scaled/normalized by row and column before running GMM? Defaults to FALSE. 
Be aware that if input modelled used rescaled data as input/option this should be set to TRUE for predicting on new data.}
}
\description{
Apply PhenoGMM made model to new data.
}
\examples{
data(flowData_transformed)
testGMM <- PhenoGMM(flowData_transformed, downsample = 1e3, nG = 128, param = c("FL1-H", "FL3-H"))
testPred <- PhenoMaskGMM(flowData_transformed, gmm = testGMM)
}
\keyword{fingerprint}
