% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_discretization.R
\name{time_discretization}
\alias{time_discretization}
\title{Time discretization function for FCM data}
\usage{
time_discretization(
  x,
  analysis.length,
  create = FALSE,
  start = 0,
  time.interval,
  height = c(0, 200),
  trigger = "FL1-H",
  time.step = 0.1
)
}
\arguments{
\item{x}{flowSet of FCM data}

\item{analysis.length}{Dataframe with $time defining the total analysis time for each sample}

\item{create}{Do you want to make a new folder for the new flowframes? Defaults to FALSE.}

\item{start}{Vector of length n(x) that indicates for each sample at what time point it should start discretisizing.
(e.g., first 10 minutes are irrelevant, start = 10*60)}

\item{time.interval}{Bin size of each new FCS file. For example, time.interval = 10 will make new FCS files of 10 second intervals.}

\item{height}{Bottom and top of the time window utilized (use minimum and maximum parameter intensity). Default on c(0,200)}

\item{trigger}{Primary parameter of signal detection (e.g. FL1-H for SYBR Green). This is used to construct the binning window with the height parameter.}

\item{time.step}{discrete unit of time used in the FCS files (FCS files don't register in seconds). This value is used to transform
the discrete time units into seconds.}
}
\description{
This function subsets the input frames based on analysis time. Developped for analysis of on-line time series data.
}
\examples{
# To be added in the near future
}
\keyword{FCM}
\keyword{analysis}
\keyword{fcm,}
\keyword{online,}
\keyword{series}
\keyword{time}
